requirements.txt
boto3
requests
beautifulsoup4
pdf2image
google-cloud-documentai
google-api-core
protobuf
torch
torchvision
torchaudio  
Pillow
opencv-python
scikit-image
matplotlib
pycocotools
camel-tools==1.5.0
tqdm
pydantic
pathlib
concurrent
urlib
time
bs4
pdf2image
boto3
logging
botocore
torch
torchvision
google.cloud
numpy
typing
google.generativeai
weaviate-client
uuid